// src/app/core/api/types/ProblemDetails.ts
export default interface ProblemDetails {
  /**
   * A URI reference that identifies the problem type. When de-referenced,
   * it provide human-readable documentation for the problem type (e.g., using HTML).
   * When this member is not present, its will return "about:blank".
   */
  type: string;
  /**
   * A short, human-readable summary of the problem type. It SHOULD NOT change from occurrence to occurrence
   * of the problem, except for purposes of localization(e.g., using proactive content negotiation;
   * see [RFC7231], Section 3.4.
   */
  title: string;
  /** The HTTP status code generated by the origin server for this occurrence of the problem. */
  status: number;
  /** A human-readable explanation specific to this occurrence of the problem. */
  detail: string;
  /** A URI reference that identifies the specific occurrence of the problem. */
  instance: string;
  /** The version of the API. */
  version: string;
  /** Timestamp in ISO format. */
  timestamp: string;
  /** Trace id. */
  traceId: string;
}